+++ CHANGES.md
@@ -1,5 +1,4 @@
-
-### 2022-12-30 version 0.9.1.1
+### 2023-09-30 version 0.9.1.1
 
 #### New features and enhancements
   * Create go wrapper for NetconfSession and other path APIs ([#779](https://github.com/CiscoDevNet/ydk-gen/issues/779))
@@ -26,6 +25,8 @@
   * CodecService fails to decode XML in python when referencing namespace prefix ([#1071](https://github.com/CiscoDevNet/ydk-gen/issues/1071))
   * Error in ydk-gen-master/test/dependencies_centos.sh version check ([#1078](https://github.com/CiscoDevNet/ydk-gen/issues/1078))
   * not_eq is a c++ keyword ([#1080](https://github.com/CiscoDevNet/ydk-gen/issues/1080))
+  * cannot decode json blob in python without encoding it first ([#1089](https://github.com/CiscoDevNet/ydk-gen/issues/1089))
+  * Incorrect bundle model reference ([#1091](https://github.com/CiscoDevNet/ydk-gen/issues/1091))
 
 ### Resolved non-GitHub issues
   * Upgraded third party packages Sphinx (1.5.6) and wheel (0.37.1) due to documentation